Partial matching & some refactoring
This is a pretty big change.

First, kana are still in lists but are converted to a dictionary for use
with romaji to hiragana conversion; this saves iterating over the whole
list each time. Mappings were also moved into their own file to separate
them from code.

Matching kanji based on partial completions is also now possible. Speed
tests indicate this is almost always fast enough it's not noticable, but
loading the dictionary takes a second or two and the very worst case
completions (just "to" or "ko" for example) can take 300ms, so that
needs work. For reference, many versions of migemo don't try to match on
two letters or less of input, probably for this reason.

There are some issues with the dictionary; nothing catastrophic but
verbs could use ending trimming for example.

One major difference from other migemos that intentional is that
optimistic kana completion, like interpeting a terminal "d" as any of
"da di du de do", is not done. It seems to mostly trigger slow
inaccurate conversions so it seems like it's better left out. -POLM