Use optimistic matching
I previously assumed this included spurious matches, but that's not
really true - using the last consonant, even uncompleted, must match
equal or fewer dictionary entries. Matching this way also reduces
flicker.

There is an issue where a hatsuon (ん) is in the middle of a word, as in
旦那, but this seems to also be an issue in other migemo
implementations. A fix isn't hard but will need a little care.

There will also be issues if the query contains brackets, but for now
this will do. -POLM